server/pyproject.toml
[project]
name = "codehood"
version = "0.1.0"
description = "An educational and social coding platform"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beartype>=0.20.2",
    "celery>=5.5.1",
    "django>=5.2",
    "django-cors-headers>=4.7.0",
    "django-katana",
    "django-model-utils>=5.0.0",
    "django-ninja>=1.4.1",
    "django-shuriken",
    "django-silk>=5.3.2",
    "django-stubs-ext>=5.1.3",
    "django-stubs[compatible-mypy]>=5.1.3",
    "django-taggit>=6.1.0",
    "django-werkzeug>=1.0.0",
    "dotenv>=0.9.9",
    "factory-boy>=3.3.3",
    "faker>=37.1.0",
    "graphene-django>=3.2.3",
    "graphene-stubs>=0.16",
    "invoke>=2.2.0",
    "markdown-it-py[plugins]>=3.0.0",
    "mdq",
    "mypy>=1.15.0",
    "pillow>=11.2.1",
    "pjrpc[openapi-ui-bundles]>=1.13.0",
    "pytest-mypy>=1.0.1",
    "rich>=14.0.0",
    "ruamel-yaml>=0.18.10",
    "rules>=3.5",
    "tomlkit>=0.13.2",
    "traitlets>=5.14.3",
    "typing-inspect>=0.9.0",
]
[project.optional-dependencies]
dev = [
    "django-types>=0.20.0",
    "django-stubs[compatible-mypy]>=5.1.3",
    "types-factory-boy>=0.4.1",
    "ipdb>=0.13.13",
    "mypy>=1.15.0",
    "rich>=14.0.0",
    "ruff>=0.11.5",
]

[project.scripts]
dev = "codehood.scripts.dev:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv.sources]
django-shuriken = { path = "../django-shuriken", editable = true }
django-katana = { path = "../django-katana", editable = true }
mdq = { path = "../mdq", editable = true }
